refactor: useHeader hook 구현 및 리팩토링
