+++ src/components/alarm/AlarmPopupItem/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ALARM } from '../../../utils/constant/alarm';
 import { extractName } from '../../../utils/validate/userList';
 import Avatar from '../../common/Avatar';
-import { AlarmItem, AlarmImg, AlarmText } from './style';
+import { AlarmImg, AlarmItem, AlarmText } from './style';
 
 const AlarmPopupItem = ({ alarm, onClick }) => {
   const { author } = alarm;
