+++ src/hooks/useHeader.js
@@ -1,0 +1,69 @@
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useRecoilState, useSetRecoilState } from 'recoil';
+import { getMyAlarms } from '../apis/alarm';
+import { isVisibleModalState } from '../recoil/addPostStates';
+import { userLoginState } from '../recoil/authState';
+import { toggleStateFamily } from '../recoil/toggleStates';
+import { TOKEN, USERIMAGE } from '../utils/constant/auth';
+import { getStorage, setStorage } from '../utils/storage';
+
+const useHeader = () => {
+  const navigate = useNavigate();
+
+  const getToken = getStorage(TOKEN);
+  const userImage = getStorage(USERIMAGE);
+
+  const [alarmBox, setAlarmBox] = useState();
+  const [alarms, setAlarms] = useState([]);
+  const setIsLogin = useSetRecoilState(userLoginState);
+  const [isAlarmOpen, setIsAlarmOpen] = useRecoilState(toggleStateFamily('alarm'));
+
+  const [isVisibleModal, setIsVisibleModal] = useRecoilState(isVisibleModalState);
+  const handleOpenAddPostModal = () => {
+    setIsVisibleModal(true);
+  };
+
+  const handleClickLogo = () => {
+    getToken ? navigate('/main') : navigate('/');
+  };
+
+  const handleOpenAlarm = async ({ target }) => {
+    if (!isAlarmOpen) {
+      setAlarmBox(target.closest('section'));
+      setIsAlarmOpen(true);
+      const response = await getMyAlarms();
+      setAlarms(response.data);
+    }
+  };
+  const handleLogout = () => {
+    setStorage(TOKEN, '');
+    setIsLogin(false);
+    navigate('/');
+  };
+
+  const handleCloseAlarm = () => {
+    setIsAlarmOpen(false);
+  };
+
+  const handleOpenMyPage = () => {
+    navigate('/myhome');
+  };
+
+  return {
+    isVisibleModal,
+    handleOpenAddPostModal,
+    handleCloseAlarm,
+    handleOpenAlarm,
+    handleLogout,
+    handleClickLogo,
+    handleOpenMyPage,
+    alarms,
+    alarmBox,
+    userImage,
+    getToken,
+    isAlarmOpen,
+  };
+};
+
+export default useHeader;
